+++ src/actions/ErrorAction.php
@@ -37,13 +37,6 @@ class ErrorAction extends \yii\web\ErrorAction
             return '';
         }
 
-        if (\Yii::$app->cms->isRedirectNotFoundHttpException && $exception instanceof NotFoundHttpException) {
-            \Yii::$app->response->redirect(Url::home());
-            \Yii::$app->response->getHeaders()->setDefault('X-CMS-REDIRECT', "isRedirectNotFoundHttpException=true");
-            \Yii::$app->end();
-            return;
-        }
-
         if ($exception instanceof \HttpException) {
             $code = $exception->statusCode;
         } else {