+++ CHANGELOG.md
@@ -17,9 +17,7 @@ CHANGELOG
 
 5.0.4
 ----------------
- * New cms option isRedirectNotFoundHttpException 
-    * true - 404 -> 302 redirrect on main site page
-    * false - 404 -> render template
+ * Fixed notices
  * Fixed create admin tree type
  * Change children elements logic
  * Cms user default username limit(1)